Afegit un gif loading en el final de la creació del cocktail Poder descartar el cocktail si es fa click en el botó back Confirmació de sortir de l'aplicació Permetre cocktails sense alcohol [WIP] Confirmació de seccions i poder descartar els elements escollits

+++ assets/www/javascripts/controllers/cocktails_controller.js
@@ -30,6 +30,8 @@ $(document).ready(function($) {
     var remain_carbonico = 1;
     var remain_vasos = 1;
     var remain_nombre = 20;
+  
+    var without_alcohol = false;
 
     var selected_data = new Object();
 
@@ -66,6 +68,90 @@ $(document).ready(function($) {
         loop : true
       });
     }
+    
+    function confirmWithoutAlcohol(button){
+      if (button == 1){
+        without_alcohol = true;
+        finish_function();
+      }
+    }
+    
+    function confirmSection(button){
+      if (button == 1){
+        finish_function();
+      }
+      else{
+        $('#finish_elem').attr("value", "Finalizar sección");
+        if (actual_product == "zumo"){
+          actual_remain = remain_zumo;
+          selected_data.zumo = new Array();
+          actual_array = selected_data.zumo;
+        } 
+        else if (actual_product == "licor"){
+          actual_remain = remain_licor;
+          $('#finish_elem').attr("value", "Cocktail sin alcohol");
+        }
+        else if (actual_product == "carbonico") actual_remain = remain_carbonico;
+        else if (actual_product == "vaso") actual_remain = remain_vaso;
+        else{
+          actual_remain = remain_nombre;
+          $('#cocktail_name').val() = "";
+        } 
+        cocktails.fillRemain();
+        mySwiper.swipeTo(0);
+      }
+    }
+    
+    function finish_function(){
+      if((actual_product == 'licor' && without_alcohol == true) || actual_array.length > 0){
+
+        //Reload page with licores data
+
+        if (actual_product == "zumo") {
+          actual_product = "licor"
+          actual_array = selected_data.licor;
+          actual_remain = remain_licores;
+          $('#title').html("LICORES")
+          $('#finish_elem').attr("value", "Cocktail sin alcohol");
+        } else if (actual_product == "licor") {
+          actual_product = "carbonico";
+          actual_array = selected_data.carbonico;
+          actual_remain = remain_carbonico;
+          $('#title').html("CARBONICOS")
+          $('#finish_elem').attr("value", "Finalizar sección");
+        } else if (actual_product == "carbonico") {
+          actual_product = "vaso";
+          actual_array = selected_data.vaso;
+          actual_remain = remain_vasos;
+          $('#title').html("VASOS");
+        } else {
+          actual_product = "nombre"
+          actual_remain = remain_nombre;
+          $('#title').html("NOMBRE");
+        }
+
+        if (actual_product != "nombre") {
+          $('.swiper-container').html("");
+          $('.swiper-container').html('<div class="swiper-wrapper"></div>');
+          cocktails.addData2Page();
+          cocktails.fillRemain();
+        } else {
+          $('.swiper-container').html("");
+         
+          $('.swiper-container').html('<input type="text" id="cocktail_name" maxlength="30"/>' + 'Te quedan <b><span id="myCounter">30</span></b> caracteres para el nombre</font>');
+          
+          var elem = $('#myCounter');
+          $('#cocktail_name').limiter(30,elem);
+          
+          $('#swiper-pagination').remove();
+          $('#remain_elem').remove();
+          $('#add_elem').remove();
+          $('#finish_elem').hide();
+          $('#buttons').append('<input type="button" id="finish_cocktail" value="¡Mezclar!" />');
+          $('#finish_cocktail').click(cocktails.finishCocktail);
+        }
+      }
+    }
 
     return {
       addData2Page : function() {
@@ -97,13 +183,14 @@ $(document).ready(function($) {
 
         for (var i = 0; i < actual_array.length; i++) {
           if (actual_array[i].id == elem.id) {
-            alert("No se puede añadir dos veces el mismo elemento");
+            navigator.notification.alert("No se puede añadir dos veces el mismo elemento", null, "Elemento repetido")
             trobat = true;
             break;
           }
         }
 
         if (!trobat) {
+          $('#finish_elem').attr("value", "Finalizar sección");
           actual_array[actual_array.length] = elem;
           actual_remain--;
           cocktails.fillRemain();
@@ -111,74 +198,46 @@ $(document).ready(function($) {
       },
       finish_element_function : function() {
         if (actual_array.length == 0) {
-          alert("Necesitas añadir como mínimo un elemento")
-        } else {
-
-          //Reload page with licores data
-
-          if (actual_product == "zumo") {
-            actual_product = "licor"
-            actual_array = selected_data.licor;
-            actual_remain = remain_licores;
-            $('#title').html("LICORES")
-          } else if (actual_product == "licor") {
-            actual_product = "carbonico";
-            actual_array = selected_data.carbonico;
-            actual_remain = remain_carbonico;
-            $('#title').html("CARBONICOS")
-          } else if (actual_product == "carbonico") {
-            actual_product = "vaso";
-            actual_array = selected_data.vaso;
-            actual_remain = remain_vasos;
-            $('#title').html("VASOS");
-          } else {
-            actual_product = "nombre"
-            actual_remain = remain_nombre;
-            $('#title').html("NOMBRE");
+          if (actual_product == "licor"){
+            navigator.notification.confirm("¿Estás seguro de querer crear un cocktail sin alcohol?", confirmWithoutAlcohol, "Cocktail sin alcohol", 'OK, Cancelar')
           }
-
-          if (actual_product != "nombre") {
-            $('.swiper-container').html("");
-            $('.swiper-container').html('<div class="swiper-wrapper"></div>');
-            cocktails.addData2Page();
-            cocktails.fillRemain();
-          } else {
-            $('.swiper-container').html("");
-           
-            $('.swiper-container').html('<input type="text" id="cocktail_name" maxlength="30"/>' + 'Te quedan <b><span id="myCounter">30</span></b> caracteres para el nombre</font>');
-            
-            var elem = $('#myCounter');
-            $('#cocktail_name').limiter(30,elem);
-            
-            $('#swiper-pagination').remove();
-            $('#remain_elem').remove();
-            $('#add_elem').remove();
-            $('#finish_elem').hide();
-            $('#buttons').append('<input type="button" id="finish_cocktail" value="¡Mezclar!" />');
-            $('#finish_cocktail').click(cocktails.finishCocktail);
+          else{
+            navigator.notification.alert("Necesitas añadir como mínimo un elemento", null, "Sección vacía")
           }
+        }
+        else{
+          var message = "Has escogido: ";
+          for (var i=0;i< actual_array.length; i++){
+            message+=("\n- " + actual_array[i].title);
+          }
+          navigator.notification.confirm(message, confirmSection, "Confirmar sección", 'OK, Cancelar')
         }
       },
       finishCocktail : function() {
-        alert("Mezclar!!!");
-        selected_data.nombre = $('#cocktail_name').val();
-
-        var json = JSON.stringify(selected_data);
-        alert(json);
-
-        //Fadeout other elements
-        
-        window.localStorage.clear();
-        
-        if(window.localStorage.getItem("n_cocktails") == null){
-          window.localStorage.setItem("n_cocktails", 0);
+        if($('#cocktail_name').val() == ''){
+          navigator.notification.alert("Debes ponerle un nombre al cocktail", null, "Nombre vacío")
         }
-        
-        window.localStorage.setItem("n_cocktails", parseInt(window.localStorage.getItem("n_cocktails")) + 1);
-
-        window.localStorage.setItem(window.localStorage.getItem("n_cocktails"), JSON.stringify(selected_data));
-        
-        loadPage('makeCocktail_final.html');
+        else{
+          alert("Mezclar!!!");
+          selected_data.nombre = $('#cocktail_name').val();
+  
+          var json = JSON.stringify(selected_data);
+          alert(json);
+  
+          //Fadeout other elements
+          
+          window.localStorage.clear();
+          
+          if(window.localStorage.getItem("n_cocktails") == null){
+            window.localStorage.setItem("n_cocktails", 0);
+          }
+          
+          window.localStorage.setItem("n_cocktails", parseInt(window.localStorage.getItem("n_cocktails")) + 1);
+  
+          window.localStorage.setItem(window.localStorage.getItem("n_cocktails"), JSON.stringify(selected_data));
+          
+          loadPage('makeCocktail_final.html');
+        }
       }
     };
   }();
